requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
psutil>=5.8.0
requests>=2.25.0
pyyaml>=6.0
click>=8.0.0
rich>=12.0.0
asyncio-mqtt>=0.11.0
prometheus-client>=0.14.0
elasticsearch>=7.15.0
redis>=4.0.0
sqlalchemy>=1.4.0
fastapi>=0.75.0
uvicorn>=0.17.0
pydantic>=1.9.0
loguru>=0.6.0